Triển khai hệ thống thông báo, mô hình quản lý kho và định nghĩa tuyến đường.

+++ ThuongMaiDienTu/database/migrations/2026_05_17_000000_create_notifications_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('notifications', function (Blueprint $table) {
             $table->bigIncrements('notification_id');
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('user_id');
             $table->string('type', 100);
             $table->string('title');
             $table->text('content');
